Use common code for rollItemMacro and Actor onSkill/PowerRoll
