+++ module/sheets/torgeternityActorSheet.js
@@ -1,4 +1,4 @@
-import { renderSkillChat } from '../torgchecks.js';
+import { renderSkillChat, rollAttack, rollPower } from '../torgchecks.js';
 import { onManageActiveEffect, prepareActiveEffectCategories } from '../effects.js';
 import { oneTestTarget, TestDialog } from '../test-dialog.js';
 import TorgeternityItem from '../documents/item/torgeternityItem.js';
@@ -803,108 +803,7 @@ export default class TorgeternityActorSheet extends foundry.applications.api.Han
   static #onAttackRoll(event, button) {
     const itemID = button.closest('.item').dataset.itemId;
     const item = this.actor.items.get(itemID);
-    const weaponData = item.system;
-    const attackWith = weaponData.attackWith;
-    let skillValue;
-    let skillData;
-    let attributes;
-
-    if (item?.weaponWithAmmo && !item.hasAmmo && !game.settings.get('torgeternity', 'ignoreAmmo')) {
-      ui.notifications.warn(game.i18n.localize('torgeternity.chatText.noAmmo'));
-      return;
-    }
-
-    if (this.actor.type === 'vehicle') {
-      skillValue = item.system.gunner.skillValue;
-      attributes = 0;
-    } else {
-      skillData = this.actor.system.skills[attackWith];
-      skillValue = skillData.value;
-      attributes = this.actor.system.attributes;
-      if (isNaN(skillValue)) {
-        skillValue = skillData.unskilledUse ? attributes[skillData.baseAttribute].value : '-';
-      }
-    }
-
-    if (checkUnskilled(skillValue, attackWith, this.actor)) return;
-
-    let dnDescriptor = 'standard';
-
-    if (game.user.targets.size) {
-      const firstTarget = game.user.targets.find(token => token.actor.type !== 'vehicle')?.actor ||
-        game.user.targets.first().actor;
-
-      if (firstTarget.type === 'vehicle') {
-        dnDescriptor = 'targetVehicleDefense';
-      } else {
-        switch (attackWith) {
-          case 'meleeWeapons':
-          case 'unarmedCombat':
-            firstTarget.items
-              .filter((it) => it.type === 'meleeweapon')
-              .filter((it) => it.system.equipped).length === 0
-              ? (dnDescriptor = 'targetUnarmedCombat')
-              : (dnDescriptor = 'targetMeleeWeapons');
-            break;
-          case 'fireCombat':
-          case 'energyWeapons':
-          case 'heavyWeapons':
-          case 'missileWeapons':
-            dnDescriptor = 'targetDodge';
-            break;
-          default:
-            dnDescriptor = 'targetMeleeWeapons';
-        }
-      }
-    }
-
-    // Calculate damage caused by this weapon
-    let adjustedDamage = 0;
-    const weaponDamage = weaponData.damage;
-    switch (weaponData.damageType) {
-      case 'flat':
-        adjustedDamage = weaponDamage;
-        break;
-      case 'strengthPlus':
-        adjustedDamage = attributes.strength.value + parseInt(weaponDamage);
-        break;
-      case 'charismaPlus':
-        adjustedDamage = attributes.charisma.value + parseInt(weaponDamage);
-        break;
-      case 'dexterityPlus':
-        adjustedDamage = attributes.dexterity.value + parseInt(weaponDamage);
-        break;
-      case 'mindPlus':
-        adjustedDamage = attributes.mind.value + parseInt(weaponDamage);
-        break;
-      case 'spiritPlus':
-        adjustedDamage = attributes.spirit.value + parseInt(weaponDamage);
-        break;
-      default:
-        adjustedDamage = parseInt(weaponDamage);
-    }
-
-    return TestDialog.wait({
-      testType: 'attack',
-      actor: this.actor,
-      amountBD: 0,
-      isAttack: true,
-      isFav: skillData?.isFav || false,
-      skillName: attackWith,
-      skillValue: Math.max(skillValue, attributes[skillData?.baseAttribute]?.value || 0),
-      unskilledUse: true,
-      damage: adjustedDamage,
-      weaponAP: weaponData.ap,
-      applyArmor: true,
-      DNDescriptor: dnDescriptor,
-      type: 'attack',
-      applySize: true,
-      attackOptions: true,
-      chatNote: weaponData.chatNote,
-      bdDamageLabelStyle: 'hidden',
-      bdDamageSum: 0,
-      itemId: item.id,
-    }, { useTargets: true });
+    rollAttack(this.actor, item);
   }
 
   /**
@@ -914,37 +813,7 @@ export default class TorgeternityActorSheet extends foundry.applications.api.Han
   static #onPowerRoll(event, button) {
     const itemID = button.closest('.item').dataset.itemId;
     const item = this.actor.items.get(itemID);
-    const powerData = item.system;
-    const skillName = powerData.skill;
-    const skillData = this.actor.system.skills[skillName];
-
-    // Set modifier for this power
-    const powerModifier = item.system.modifier || 0;
-
-    if (checkUnskilled(skillData.value, skillName, this.actor)) return;
-
-    return TestDialog.wait({
-      testType: 'power',
-      actor: this.actor,
-      powerName: item.name,
-      powerModifier: powerModifier,
-      isAttack: powerData.isAttack,
-      isFav: skillData.isFav,
-      skillName: skillName,
-      skillAdds: skillData.adds,
-      skillValue: Math.max(skillData.value, this.actor.system.attributes[skillData.baseAttribute].value),
-      unskilledUse: false,
-      damage: powerData.damage,
-      weaponAP: powerData.ap,
-      applyArmor: powerData.applyArmor,
-      DNDescriptor: powerData.dn,
-      applySize: powerData.applySize,
-      attackOptions: true,
-      bdDamageLabelStyle: 'dihiddene',
-      amountBD: 0,
-      bdDamageSum: 0,
-      itemId: item.id,
-    }, { useTargets: true });
+    rollPower(this.actor, item);
   }
 
   /**
